LayoutTests/platform/gtk/accessibility/xml-roles-exposed-expected.txt


Verify exposure of ARIA role values via the 'xml-roles' object attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: AXAlert
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:alert, xml-roles:alert, toolkit:WebKitGtk
------------
AXRole: AXAlert
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:alertdialog, xml-roles:alertdialog, toolkit:WebKitGtk
------------
AXRole: AXEmbedded
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:application, xml-roles:application, toolkit:WebKitGtk
------------
AXRole: AXArticle
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:article, xml-roles:article, toolkit:WebKitGtk
------------
AXRole: AXLandmarkBanner
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:banner, xml-roles:banner, toolkit:WebKitGtk
------------
AXRole: AXButton
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:button, xml-roles:button, toolkit:WebKitGtk
------------
AXRole: AXCheckBox
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:checkbox, xml-roles:checkbox, toolkit:WebKitGtk
------------
AXRole: AXComboBox
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:combobox, xml-roles:combobox, toolkit:WebKitGtk
------------
AXRole: AXLandmarkComplementary
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:complementary, xml-roles:complementary, toolkit:WebKitGtk
------------
AXRole: AXLandmarkContentInfo
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:contentinfo, xml-roles:contentinfo, toolkit:WebKitGtk
------------
AXRole: AXDefinition
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:definition, xml-roles:definition, toolkit:WebKitGtk
------------
AXRole: AXDialog
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:dialog, xml-roles:dialog, toolkit:WebKitGtk
------------
AXRole: AXList
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:list, xml-roles:list, toolkit:WebKitGtk
------------
AXRole: AXDocument
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:document, xml-roles:document, toolkit:WebKitGtk
------------
AXRole: AXForm
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:form, xml-roles:form, toolkit:WebKitGtk
------------
AXRole: AXGroup
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:group, xml-roles:group, toolkit:WebKitGtk
------------
AXRole: AXHeading
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:heading, xml-roles:heading, toolkit:WebKitGtk
------------
AXRole: AXImage
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:img, xml-roles:img, toolkit:WebKitGtk
------------
AXRole: AXLink
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXURL: (null)
AXPlatformAttributes: computed-role:link, xml-roles:link, toolkit:WebKitGtk
------------
AXRole: AXLog
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:log, xml-roles:log, toolkit:WebKitGtk
------------
AXRole: AXLandmarkMain
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:main, xml-roles:main, toolkit:WebKitGtk
------------
AXRole: AXMarquee
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:marquee, xml-roles:marquee, toolkit:WebKitGtk
------------
AXRole: AXMath
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:math, xml-roles:math, toolkit:WebKitGtk
------------
AXRole: AXLandmarkNavigation
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:navigation, xml-roles:navigation, toolkit:WebKitGtk
------------
AXRole: AXComment
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:note, xml-roles:note, toolkit:WebKitGtk
------------
AXRole: AXProgressIndicator
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:progressbar, xml-roles:progressbar, toolkit:WebKitGtk
------------
AXRole: AXRadioButton
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 1
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:radio, xml-roles:radio, toolkit:WebKitGtk
------------
AXRole: AXGroup
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:region, xml-roles:region, toolkit:WebKitGtk
------------
AXRole: AXScrollBar
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:scrollbar, xml-roles:scrollbar, toolkit:WebKitGtk
------------
AXRole: AXLandmarkSearch
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:search, xml-roles:search, toolkit:WebKitGtk
------------
AXRole: AXTextField
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:searchbox, xml-roles:searchbox, toolkit:WebKitGtk
------------
AXRole: AXSeparator
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:separator, xml-roles:separator, toolkit:WebKitGtk
------------
AXRole: AXSlider
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:slider, xml-roles:slider, toolkit:WebKitGtk
------------
AXRole: AXSpinButton
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:spinbutton, xml-roles:spinbutton, toolkit:WebKitGtk
------------
AXRole: AXStatusBar
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:status, xml-roles:status, toolkit:WebKitGtk
------------
AXRole: AXToggleButton
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:switch, xml-roles:switch, toolkit:WebKitGtk
------------
AXRole: AXTextField
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: toolkit:WebKitGtk
------------
AXRole: AXTimer
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:timer, xml-roles:timer, toolkit:WebKitGtk
------------
AXRole: AXToolbar
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:toolbar, xml-roles:toolbar, toolkit:WebKitGtk
------------
AXRole: AXUserInterfaceTooltip
AXParent: AXWebArea
AXChildren: 0
AXPosition: { 0.000000, 0.000000 }
AXSize: { 0.000000, 0.000000 }
AXTitle: 
AXDescription: 
AXValue: 
AXFocusable: 0
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: computed-role:tooltip, xml-roles:tooltip, toolkit:WebKitGtk
------------
AXRole: AXParagraph
AXParent: AXWebArea
AXChildren: 2
AXPosition: { 8.000000, 8.000000 }
AXSize: { 784.000000, 50.000000 }
AXTitle: 
AXDescription: 
AXValue: Verify exposure of ARIA role values via the 'xml-roles' object attribute.<\n><\n>On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".<\n><\n><\n>
AXFocusable: 1
AXFocused: 0
AXSelectable: 0
AXSelected: 0
AXMultiSelectable: 0
AXEnabled: 1
AXExpanded: 0
AXRequired: 0
AXChecked: 0
AXPlatformAttributes: html-id:description, toolkit:WebKitGtk
------------

PASS successfullyParsed is true

TEST COMPLETE
